+++ src/Phinx/Db/Adapter/MysqlAdapter.php
@@ -905,7 +905,7 @@ class MysqlAdapter extends PdoAdapter
      *
      * @throws \Phinx\Db\Adapter\UnsupportedColumnTypeException
      */
-    public function getSqlType($type, ?int $limit = null): array
+    public function getSqlType($type, ?float $limit = null): array
     {
         switch ($type) {
             case static::PHINX_TYPE_FLOAT:
